Renamed request data class to Message dataclass, deleted token from connection dataclass
